add os path handler

+++ python-server_side_rendering/task_00_intro.py
@@ -2,11 +2,12 @@
 """ function to replace words in templates """
 
 import logging
+import os
 
 
 def generate_invitations(template, attendees):
     """ function to replace templates with a list of dicts"""
-    # Input checking types
+    # Input checking - types
     if not isinstance(template, str):
         logging.error("Tamplate should be a string")
         raise ValueError("Not a string")
@@ -15,7 +16,7 @@ def generate_invitations(template, attendees):
         logging.error("Attendees should be a list of dictionaries")
         raise ValueError("Not a list")
 
-    # Empty inputs
+    # check empty inputs
     if not template:
         logging.error("Tamplate is empty, no output files generated.")
         raise IndexError("Tamplate is empty, no output files generated.")
@@ -25,6 +26,7 @@ def generate_invitations(template, attendees):
         raise IndexError("No data provided, no output files generated.")
 
 
+    # Start index from 1
     for index, attendee in enumerate(attendees, start=1):
         output_content = template
         for key, value in attendee.items():
@@ -34,7 +36,12 @@ def generate_invitations(template, attendees):
 
         # Output file
         # name = attendee.get("name", f"attendee_{index}").replace(" ", "_")
-        output_filename = f"output_{index + 1}.txt"
+        output_filename = f"output_{index}.txt"
+
+        # check if file already exists
+        if os.path.exists(output_filename):
+            logging.warning("File %s already exists, skipping.", output_filename)
+            continue
 
         try:
             with open(output_filename, "w", encoding="utf-8") as output_file:
